# src/assets/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LoggerMessages.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/EventGroups": {
      "get": {
        "tags": [
          "EventGroups"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventGroup"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventGroup"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventGroup"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EventGroups"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventGroups/{id}": {
      "get": {
        "tags": [
          "EventGroups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EventGroups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "EventGroups"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventGroups/Find/{abbr}_{slnId}_{slnName}": {
      "get": {
        "tags": [
          "EventGroups"
        ],
        "parameters": [
          {
            "name": "abbr",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slnId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "slnName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroup"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventGroup": {
        "type": "object",
        "properties": {
          "oid": {
            "type": "string",
            "format": "uuid"
          },
          "solutionId": {
            "type": "string",
            "format": "uuid"
          },
          "solutionName": {
            "type": "string",
            "nullable": true
          },
          "eventGroupAbbr": {
            "maxLength": 3,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}